+++ delve.c
@@ -325,7 +325,7 @@ save_delve()
 	struct character *curchar = get_current_character();
 	char path[_POSIX_PATH_MAX];
 	json_object *root, *items, *id;
-	int temp_n, i;
+	size_t temp_n, i;
 
 	if (curchar == NULL) {
 		log_debug("No character loaded.  No delve to save.\n");
@@ -391,7 +391,7 @@ delete_delve(int id)
 {
 	char path[_POSIX_PATH_MAX];
 	json_object *root, *lid;
-	int temp_n, i;
+	size_t temp_n, i;
 
 	snprintf(path, sizeof(path), "%s/delve.json", get_isscrolls_dir());
 	if ((root = json_object_from_file(path)) == NULL) {
@@ -429,7 +429,7 @@ load_delve(int id)
 	struct character *curchar = get_current_character();
 	char path[_POSIX_PATH_MAX];
 	json_object *root, *lid;
-	int temp_n, i;
+	size_t temp_n, i;
 
 	if (curchar == NULL) {
 		log_debug("No character loaded\n");
