+++ fight.c
@@ -373,7 +373,7 @@ save_fight()
 	struct character *curchar = get_current_character();
 	char path[_POSIX_PATH_MAX];
 	json_object *root, *items, *id;
-	int temp_n, i;
+	size_t temp_n, i;
 
 	if (curchar == NULL) {
 		log_debug("No character loaded.  No fight to save.\n");
@@ -438,7 +438,7 @@ delete_fight(int id)
 {
 	char path[_POSIX_PATH_MAX];
 	json_object *root, *lid;
-	int temp_n, i;
+	size_t temp_n, i;
 
 	snprintf(path, sizeof(path), "%s/fight.json", get_isscrolls_dir());
 	if ((root = json_object_from_file(path)) == NULL) {
@@ -476,7 +476,7 @@ load_fight(int id)
 	struct character *curchar = get_current_character();
 	char path[_POSIX_PATH_MAX];
 	json_object *root, *lid;
-	int temp_n, i;
+	size_t temp_n, i;
 
 	if (curchar == NULL) {
 		log_debug("No character loaded\n");
